move recoil atom to a separate direcotry
