+++ src/components/asset-list/RecyclerAssetList2/profile-header/ProfileActionButtonsRow.tsx
@@ -34,10 +34,10 @@ import showWalletErrorAlert from '@/helpers/support';
 import { analytics, analyticsV2 } from '@/analytics';
 import ContextMenuButton from '@/components/native-context-menu/contextMenu';
 import { useRecoilState } from 'recoil';
-import { addressCopiedToastAtom } from '@/screens/WalletScreen';
 import config from '@/model/config';
 import { useAccountAccentColor } from '@/hooks/useAccountAccentColor';
 import { getAllActiveSessionsSync } from '@/utils/walletConnect';
+import { addressCopiedToastAtom } from '@/recoil/addressCopiedToastAtom';
 
 export const ProfileActionButtonsRowHeight = 80;
