+++ src/recoil/addressCopiedToastAtom.tsx
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const addressCopiedToastAtom = atom({
+  default: false,
+  key: 'addressCopiedToast',
+});
